Time: 7 ms (90.42%), Space: 13.7 MB (63.15%) - LeetHub

+++ 34-find-first-and-last-position-of-element-in-sorted-array/34-find-first-and-last-position-of-element-in-sorted-array.cpp
@@ -5,8 +5,8 @@ public:
         vector<int> :: iterator lo, hi;
         lo = lower_bound(nums.begin(), nums.end(), target);
         hi = upper_bound(nums.begin(), nums.end(), target);
-        int loindex = (int)(lo - nums.begin());
-        int hiindex = (int)(hi - nums.begin());
+        int loindex = lo - nums.begin();
+        int hiindex = hi - nums.begin();
         if (lo == nums.end() || *lo != target) return {-1, -1};
         return {loindex, hiindex - 1};
     }
